feat(stock): support fractional quantities and unit types

+++ src/main/java/com/eduribeiro8/LilMarket/dto/BatchLossReportRequestDTO.java
@@ -5,12 +5,14 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
+
 @Schema(description = "Requisição para reportar perda (danos/estrago) de um lote")
 public record BatchLossReportRequestDTO(
         @Schema(description = "Identificador do lote", example = "7", requiredMode = Schema.RequiredMode.REQUIRED)
         @NotNull Integer batchId,
         @Schema(description = "Quantidade perdida", example = "2", requiredMode = Schema.RequiredMode.REQUIRED)
-        @Positive int quantity,
+        @Positive BigDecimal quantity,
         @Schema(description = "Motivo da perda", example = "Quebrado durante transporte", requiredMode = Schema.RequiredMode.REQUIRED)
         @NotBlank String reason
 ) {
